+++ components/popups/PopupFrameEditCurrency.vue
@@ -2,9 +2,15 @@
   <div class="frame">
     <h1 v-text="`Edit ${data.data.name}`" />
     <Layout v-if="currency" name="component-flow">
-      <Input v-model="currency.id" label="General - Id" type="text" :disabled="true" />
-      <Input v-model="currency.name" label="General - Name" type="text" />
-      <Input v-model="currency.symbol" label="General - Symbol" type="text" />
+      <h3>General</h3>
+      <Layout name="2static">
+        <Input v-model="currency.id" label="Id" type="text" :disabled="true" />
+        <Input v-model="currency.code" label="Code" type="text" placeholder="eur" />
+      </Layout>
+      <Layout name="2static">
+        <Input v-model="currency.name" label="Name" type="text" placeholder="Euro" />
+        <Input v-model="currency.symbol" label="Symbol" type="text" placeholder="€" />
+      </Layout>
 
       <Admonition v-if="error" type="error" :text="error" />
 
@@ -74,6 +80,11 @@ export default Vue.extend({
       this.error = ''
     },
     async save() {
+      if (!/^[a-z]{3}$/.test(this.currency.code)) {
+        this.error = 'Code must match /^[a-z]{3}$/'
+        return
+      }
+
       const { data, status, statusText } = await API.patchCurrency(this.data.data.id, this.currency)
       if (status !== 200) {
         this.error = `Error. http ${status}: ${statusText}.\n${data?.error}: ${data?.message}`
