+++ pages/apps/home.vue
@@ -1,24 +1,32 @@
 <template>
-  <div class="container">
+  <Container>
     <span v-if="!data">Loading...</span>
     <div v-else-if="data.none">
       <h1>Welcome!</h1>
-      <span>Looks like you haven't created an application yet.</span>
+      <p>Looks like you haven't created an application yet.</p>
       <br>
-      <span>Click below to get started with using the FreeStuff API!</span>
+      <p>Click below to get started with using the FreeStuff API!</p>
       <br>
       <br>
-      <button generic dark @click="createApp">
-        Create Application
-      </button>
+      <Button
+        type="green"
+        text="Create Application"
+        @click="createApp"
+      />
     </div>
     <div v-else>
-      <h1 v-text="data.type == 'partner' ? 'Your application [partner]:' : 'Your application:'" />
-      <h2>About your app:</h2>
-      <span>Please describe your app in a few words. What are you using it for? Can you link some websites / social media for your project?</span>
-      <textarea v-model="newdesc" maxlength="2048" @blur="saveDesc" />
+      <h1 v-text="data.type == 'partner' ? 'Your application [partner]' : 'Your application'" />
+      <h2>About your app</h2>
+      <p>Please describe your app in a few words. What are you using it for? Can you link some websites / social media for your project?</p>
+      <Input
+        v-model="newdesc"
+        label="Description"
+        :multiline="true"
+        @blur="saveDesc"
+      />
 
-      <h2><Icon name="key" /> API Key:</h2>
+      <!-- <h2><Icon name="key" /> API Key</h2> -->
+      <h2>API Key</h2>
       <div class="content-row">
         <div class="tokenbox">
           <span class="curtain">Hover to view</span>
@@ -30,51 +38,49 @@
       </div>
       <sub>Last regenerated {{ getLastChanged('key') }}</sub>
 
-      <h2><Icon name="webhook" /> Event Webhook:</h2>
-      <span>General Settings</span>
-      <div class="spacer small" />
-      <div class="content-row">
-        <input
+      <!-- <h2><Icon name="webhook" /> Event Webhook</h2> -->
+      <h2>Event Webhooks</h2>
+      <!-- <div class="spacer small" /> -->
+      <Layout name="component-flow">
+        <Input
           v-model="data.webhook"
-          v-tippy="{ placement: 'top' }"
           :disabled="!webhookEdit"
-          type="text"
+          label="Webhook URL"
           placeholder="https://your-domain.com/webhooks/freestuff"
-          content="Webhook URL"
-        >
-      </div>
-      <div class="content-row">
-        <input
+        />
+        <Input
           v-model="data.webhooksecret"
-          v-tippy="{ placement: 'top' }"
           :disabled="!webhookEdit"
-          type="text"
-          placeholder="Webhook Secret"
-          content="Webhook Secret"
-        >
-        <button generic dark style="margin-right: 10pt;" @click="updateWebhook">
-          {{ webhookEdit ? 'Save Changes' : 'Edit' }}
-        </button>
-        <button generic dark @click="testWebhook">
-          {{ webhookEdit ? 'Cancel' : 'Test Webhook' }}
-        </button>
-      </div>
-      <sub>Last changed {{ getLastChanged('webhook') }}</sub>
+          label="Webhook Secret"
+          placeholder="Secret"
+        />
+        <!-- <sub>Last changed {{ getLastChanged('webhook') }}</sub> -->
 
-      <div class="spacer medium" />
-      <span>Webhook Version</span>
-      <div class="spacer small" />
-      <div class="content-row">
-        <select v-model="data.webhookVersion">
-          <option value="1">
-            Version 1 &mdash; Default
-          </option>
-          <!-- <option value="2">
-            Version 2 &mdash; NEW
-          </option> -->
-        </select>
-      </div>
-      <sub>Last changed {{ getLastChanged('webhookVersion') }}</sub>
+        <Layout name="inline">
+          <Button
+            type="light"
+            :text="webhookEdit ? 'Save Changes' : 'Edit'"
+            @click="updateWebhook"
+          />
+          <Button
+            type="light"
+            :text="webhookEdit ? 'Cancel' : 'Test Webhook'"
+            @click="testWebhook"
+          />
+        </Layout>
+
+        <div class="spacer small" />
+
+        <Input
+          v-model="data.webhookVersion"
+          label="Webhook Version"
+          :options="[
+            { label: 'Version 1 — Default', value: '1' }
+          ]"
+        />
+      <!-- <sub>Last changed {{ getLastChanged('webhookVersion') }}</sub> -->
+      </Layout>
+
 
       <div v-if="data.webhookVersion == 2">
         <div class="spacer medium" />
@@ -83,19 +89,30 @@
         <span>TODO</span>
       </div>
 
-      <h2><Icon name="analytics" /> API Usage:</h2>
+      <!-- <h2><Icon name="analytics" /> API Usage</h2> -->
+      <h2>API Usage</h2>
       <span>Coming soon&trade;</span>
     </div>
-  </div>
+  </Container>
 </template>
 
 <script lang="ts">
 import Vue from 'vue'
 import Swal from 'sweetalert2'
 import API from '../../lib/api'
+import Container from '~/components/layout/Container.vue'
+import Layout from '~/components/layout/Layout.vue'
+import Button from '~/components/entities/Button.vue'
+import Input from '~/components/entities/Input.vue'
 
 
 export default Vue.extend({
+  components: {
+    Container,
+    Layout,
+    Button,
+    Input
+  },
   transition: 'slide-down',
   async fetch() {
     const info = await API.getAppData()
