+++ store/lang.js
@@ -3,11 +3,10 @@ export const state = () => ({
   declined: '✕ Declined',
   approved: '✓ Approved',
   processing: '… Processing',
-  free: 'Free to Keep',
-  weekend: 'Free for a Weekend',
-  discount: 'Discount',
-  ad: 'Advertisement',
-  unknown: 'Unknown',
+  keep: 'Free to Keep',
+  timed: 'Free for a Weekend',
+  prime: 'Included in Prime',
+  gamepass: 'Included in GamePass',
   steam: 'Steam',
   epic: 'Epic Games Store',
   humble: 'Humble Bundle',
